Add files via upload
fixed sensitivity analysis (sort of). epsilon is zero for all models for all equations, yet it should be nonzero for dR/dt and dS/dt. 

notes:
"
For No Quarantine, No Vaccination Model: beta and epsilon should be nonzero for dS/dt; beta, gamma, and delta should be nonzero for dI/dt; gamma and epsilon should be nonzero for dR/dt; delta should be nonzero for dDdt.

For Quarantine, No Vaccination Model: beta, epsilon, and q should be nonzero for dS/dt; beta, gamma, delta, and q should be nonzero for dI/dt; gamma and epsilon should be nonzero for dR/dt; delta should be nonzero for dDdt.

For No Quarantine, With Vaccination Model: beta, epsilon, and v should be nonzero for dS/dt; beta, gamma, and delta should be nonzero for dI/dt; gamma, epsilon, and v should be nonzero for dR/dt; delta should be nonzero for dDdt.

For Quarantine and Vaccination Model: beta, epsilon, v, and q should be nonzero for dS/dt; beta, gamma, delta, and q should be nonzero for dI/dt; gamma, epsilon, and v should be nonzero for dR/dt; delta should be nonzero for dDdt.
"